fix(backend): Move ginx file to admin. issue: #15717
